Revert changes in integration tests
